DT-727 Split unit tests by config manager type
