+++ src/test/kotlin/uk/gov/justice/digital/hmpps/gradle/pluginmanagers/DependencyManagementPluginManagerTest.kt
@@ -1,0 +1,24 @@
+package uk.gov.justice.digital.hmpps.gradle.pluginmanagers
+
+import io.spring.gradle.dependencymanagement.DependencyManagementPlugin
+import org.gradle.api.Project
+import org.gradle.testfixtures.ProjectBuilder
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+class DependencyManagementPluginManagerTest {
+
+  private val project: Project = ProjectBuilder.builder().build()
+
+  @BeforeEach
+  fun `Create project`() {
+    project.pluginManager.apply("uk.gov.justice.digital.hmpps.gradle.DpsSpringBoot")
+  }
+
+  @Test
+  fun `Should apply the Spring Dependency Management plugin`() {
+    project.plugins.getPlugin(DependencyManagementPlugin::class.java)
+  }
+
+
+}
